primer commit

+++ app/Http/Controllers/OrdenDeTrabajoController.php
@@ -19,7 +19,7 @@ class OrdenDeTrabajoController extends Controller
      */
     public function index()
     {
-        $ordentrabajo = orden_de_trabajo::with('equipo_de_trabajo','cliente')->paginate(2);
+        $ordentrabajo = orden_de_trabajo::with('equipo_de_trabajo','cliente')->paginate(3);
         return view('Orden.index', ['ordentrabajo' => $ordentrabajo]);
 
     }
@@ -150,18 +150,20 @@ class OrdenDeTrabajoController extends Controller
      * @param  \App\Models\orden_de_trabajo  $orden_de_trabajo
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
-    {
-        $orden_de_trabajo = orden_de_trabajo::find($id);
-        $cliente = Cliente::findOrFail($orden_de_trabajo->cliente_id);
-        $orden_de_trabajo -> delete();
-        $cliente_ordenes = orden_de_trabajo::where('cliente_id', $cliente->id)->count();
 
-        if ($cliente_ordenes == 0) {
+     public function destroy($id)
+     {
+         $orden_de_trabajo = orden_de_trabajo::find($id);
+         if ($orden_de_trabajo) {
+             $cliente = Cliente::findOrFail($orden_de_trabajo->cliente_id);
+             $orden_de_trabajo->delete();
+             $cliente_ordenes = orden_de_trabajo::where('cliente_id', $cliente->id)->count();
+             if ($cliente_ordenes == 0) {
+                 $cliente->delete();
+             }
+             return redirect("orden_de_trabajo/")->with('success', 'Orden eliminada exitosamente.');
+         }
+         return redirect("orden_de_trabajo/")->with('error', 'Orden no encontrada.');
+     }
 
-            $cliente->delete();
-    }
-
-        return redirect("orden_de_trabajo/");
-    }
 }
